+++ packages/server/graphql/mutations/endCheckIn.ts
@@ -7,13 +7,13 @@ import {positionAfter} from '../../../client/shared/sortOrder'
 import {checkTeamsLimit} from '../../billing/helpers/teamLimitsCheck'
 import getRethink from '../../database/rethinkDriver'
 import {RDatum} from '../../database/stricterR'
-import MeetingAction from '../../database/types/MeetingAction'
 import Task from '../../database/types/Task'
 import TimelineEventCheckinComplete from '../../database/types/TimelineEventCheckinComplete'
 import {DataLoaderInstance} from '../../dataloader/RootDataLoader'
 import generateUID from '../../generateUID'
 import getKysely from '../../postgres/getKysely'
 import {AgendaItem} from '../../postgres/types'
+import {CheckInMeeting} from '../../postgres/types/Meeting'
 import archiveTasksForDB from '../../safeMutations/archiveTasksForDB'
 import removeSuggestedAction from '../../safeMutations/removeSuggestedAction'
 import {Logger} from '../../utils/Logger'
@@ -81,6 +81,7 @@ const clonePinnedAgendaItems = async (
   pinnedAgendaItems: AgendaItem[],
   dataLoader: DataLoaderInstance
 ) => {
+  if (!pinnedAgendaItems.length) return
   let curSortOrder = ''
   const clonedPins = pinnedAgendaItems.map((agendaItem) => {
     const agendaItemId = `${agendaItem.teamId}::${generateUID()}`
@@ -100,12 +101,13 @@ const clonePinnedAgendaItems = async (
   dataLoader.clearAll('agendaItems')
 }
 
-const summarizeCheckInMeeting = async (meeting: MeetingAction, dataLoader: DataLoaderWorker) => {
+const summarizeCheckInMeeting = async (meeting: CheckInMeeting, dataLoader: DataLoaderWorker) => {
   /* If isKill, no agenda items were processed so clear none of them.
    * Similarly, don't clone pins. the original ones will show up again.
    */
 
   const {id: meetingId, teamId, phases} = meeting
+  const pg = getKysely()
   const r = await getRethink()
   const [meetingMembers, tasks, doneTasks, activeAgendaItems] = await Promise.all([
     dataLoader.get('meetingMembersByMeetingId').load(meetingId),
@@ -141,6 +143,15 @@ const summarizeCheckInMeeting = async (meeting: MeetingAction, dataLoader: DataL
     isKill ? undefined : archiveTasksForDB(doneTasks, meetingId),
     isKill ? undefined : clonePinnedAgendaItems(pinnedAgendaItems, dataLoader),
     updateTaskSortOrders(userIds, tasks),
+    pg
+      .updateTable('NewMeeting')
+      .set({
+        agendaItemCount: activeAgendaItems.length,
+        commentCount,
+        taskCount: tasks.length
+      })
+      .where('id', '=', meetingId)
+      .execute(),
     r
       .table('NewMeeting')
       .get(meetingId)
@@ -154,7 +165,7 @@ const summarizeCheckInMeeting = async (meeting: MeetingAction, dataLoader: DataL
       )
       .run()
   ])
-
+  dataLoader.clearAll('newMeetings')
   return {updatedTaskIds: [...tasks, ...doneTasks].map(({id}) => id)}
 }
 
@@ -169,6 +180,7 @@ export default {
   },
   async resolve(_source: unknown, {meetingId}: {meetingId: string}, context: GQLContext) {
     const {authToken, socketId: mutatorId, dataLoader} = context
+    const pg = getKysely()
     const r = await getRethink()
     const operationId = dataLoader.share()
     const subOptions = {mutatorId, operationId}
@@ -176,12 +188,11 @@ export default {
     const viewerId = getUserId(authToken)
 
     // AUTH
-    const meeting = (await r
-      .table('NewMeeting')
-      .get(meetingId)
-      .default(null)
-      .run()) as MeetingAction | null
+    const meeting = await dataLoader.get('newMeetings').load(meetingId)
     if (!meeting) return standardError(new Error('Meeting not found'), {userId: viewerId})
+    if (meeting.meetingType !== 'action') {
+      return standardError(new Error('Not a check-in meeting'), {userId: viewerId})
+    }
     const {endedAt, facilitatorStageId, phases, teamId} = meeting
 
     // VALIDATION
@@ -200,7 +211,7 @@ export default {
     const phase = getMeetingPhase(phases)
     const insights = await gatherInsights(meeting, dataLoader)
 
-    const completedCheckIn = (await r
+    const completedCheckIn = await r
       .table('NewMeeting')
       .get(meetingId)
       .update(
@@ -212,14 +223,29 @@ export default {
         {returnChanges: true}
       )('changes')(0)('new_val')
       .default(null)
-      .run()) as unknown as MeetingAction
-
+      .run()
+    await pg
+      .updateTable('NewMeeting')
+      .set({
+        endedAt: now,
+        phases: JSON.stringify(phases),
+        usedReactjis: JSON.stringify(insights.usedReactjis),
+        engagement: insights.engagement
+      })
+      .where('id', '=', meetingId)
+      .execute()
+    dataLoader.clearAll('newMeetings')
     if (!completedCheckIn) {
       return standardError(new Error('Completed check-in meeting does not exist'), {
         userId: viewerId
       })
     }
 
+    if (completedCheckIn.meetingType !== 'action') {
+      return standardError(new Error('Completed check-in meeting is not an action'), {
+        userId: viewerId
+      })
+    }
     // remove any empty tasks
     const [meetingMembers, team, teamMembers, removedTaskIds] = await Promise.all([
       dataLoader.get('meetingMembersByMeetingId').load(meetingId),
@@ -246,7 +272,6 @@ export default {
         })
     )
     const timelineEventId = events[0]!.id
-    const pg = getKysely()
     await pg.insertInto('TimelineEvent').values(events).execute()
     if (team.isOnboardTeam) {
       const teamMembers = await dataLoader.get('teamMembersByTeamId').load(teamId)