+++ packages/server/graphql/mutations/endSprintPoker.ts
@@ -1,11 +1,10 @@
 import {GraphQLID, GraphQLNonNull} from 'graphql'
+import {sql} from 'kysely'
 import {SubscriptionChannel} from 'parabol-client/types/constEnums'
 import getMeetingPhase from 'parabol-client/utils/getMeetingPhase'
 import findStageById from 'parabol-client/utils/meetings/findStageById'
 import {checkTeamsLimit} from '../../billing/helpers/teamLimitsCheck'
 import getRethink from '../../database/rethinkDriver'
-import Meeting from '../../database/types/Meeting'
-import MeetingPoker from '../../database/types/MeetingPoker'
 import TimelineEventPokerComplete from '../../database/types/TimelineEventPokerComplete'
 import getKysely from '../../postgres/getKysely'
 import {Logger} from '../../utils/Logger'
@@ -41,12 +40,11 @@ export default {
     const viewerId = getUserId(authToken)
 
     // AUTH
-    const meeting = (await r
-      .table('NewMeeting')
-      .get(meetingId)
-      .default(null)
-      .run()) as Meeting | null
+    const meeting = await dataLoader.get('newMeetings').load(meetingId)
     if (!meeting) return standardError(new Error('Meeting not found'), {userId: viewerId})
+    if (meeting.meetingType !== 'poker') {
+      return standardError(new Error('Meeting is not a poker meeting'), {userId: viewerId})
+    }
     const {endedAt, facilitatorStageId, phases, teamId} = meeting
 
     // VALIDATION
@@ -81,7 +79,7 @@ export default {
       await dataLoader.get('commentCountByDiscussionId').loadMany(discussionIds)
     ).filter(isValid)
     const commentCount = commentCounts.reduce((cumSum, count) => cumSum + count, 0)
-    const completedMeeting = (await r
+    const completedMeeting = await r
       .table('NewMeeting')
       .get(meetingId)
       .update(
@@ -95,18 +93,34 @@ export default {
         {returnChanges: true, nonAtomic: true}
       )('changes')(0)('new_val')
       .default(null)
-      .run()) as unknown as MeetingPoker
+      .run()
+    await getKysely()
+      .updateTable('NewMeeting')
+      .set({
+        endedAt: sql`CURRENT_TIMESTAMP`,
+        phases: JSON.stringify(phases),
+        commentCount,
+        storyCount,
+        usedReactjis: JSON.stringify(insights.usedReactjis),
+        engagement: insights.engagement
+      })
+      .where('id', '=', meetingId)
+      .executeTakeFirst()
+    dataLoader.clearAll('newMeetings')
     if (!completedMeeting) {
       return standardError(new Error('Completed poker meeting does not exist'), {
         userId: viewerId
       })
+    }
+    if (completedMeeting.meetingType !== 'poker') {
+      return standardError(new Error('Meeting is not a poker meeting'), {userId: viewerId})
     }
     const {templateId} = completedMeeting
     const [meetingMembers, team, teamMembers, removedTaskIds, template] = await Promise.all([
       dataLoader.get('meetingMembersByMeetingId').load(meetingId),
       dataLoader.get('teams').loadNonNull(teamId),
       dataLoader.get('teamMembersByTeamId').load(teamId),
-      removeEmptyTasks(meetingId),
+      removeEmptyTasks(meetingId, teamId),
       // technically, this template could have mutated while the meeting was going on. but in practice, probably not
       dataLoader.get('meetingTemplates').loadNonNull(templateId)
     ])