+++ packages/server/graphql/mutations/helpers/safeEndTeamPrompt.ts
@@ -1,10 +1,11 @@
+import {sql} from 'kysely'
 import {SubscriptionChannel} from 'parabol-client/types/constEnums'
 import {checkTeamsLimit} from '../../../billing/helpers/teamLimitsCheck'
 import getRethink, {ParabolR} from '../../../database/rethinkDriver'
-import MeetingTeamPrompt from '../../../database/types/MeetingTeamPrompt'
 import TimelineEventTeamPromptComplete from '../../../database/types/TimelineEventTeamPromptComplete'
 import getKysely from '../../../postgres/getKysely'
 import {getTeamPromptResponsesByMeetingId} from '../../../postgres/queries/getTeamPromptResponsesByMeetingIds'
+import {TeamPromptMeeting} from '../../../postgres/types/Meeting'
 import {Logger} from '../../../utils/Logger'
 import {analytics} from '../../../utils/analytics/analytics'
 import publish, {SubOptions} from '../../../utils/publish'
@@ -16,12 +17,15 @@ import generateStandupMeetingSummary from './generateStandupMeetingSummary'
 import {IntegrationNotifier} from './notifications/IntegrationNotifier'
 import updateQualAIMeetingsCount from './updateQualAIMeetingsCount'
 
-const summarizeTeamPrompt = async (meeting: MeetingTeamPrompt, context: InternalContext) => {
+const summarizeTeamPrompt = async (meeting: TeamPromptMeeting, context: InternalContext) => {
   const {dataLoader} = context
+  const pg = getKysely()
   const r = await getRethink()
 
   const summary = await generateStandupMeetingSummary(meeting, dataLoader)
-
+  if (summary) {
+    await pg.updateTable('NewMeeting').set({summary}).where('id', '=', meeting.id).execute()
+  }
   await r
     .table('NewMeeting')
     .get(meeting.id)
@@ -30,7 +34,7 @@ const summarizeTeamPrompt = async (meeting: MeetingTeamPrompt, context: Internal
     })
     .run()
 
-  dataLoader.get('newMeetings').clear(meeting.id)
+  dataLoader.clearAll('newMeetings')
   // wait for whole meeting summary to be generated before sending summary email and updating qualAIMeetingCount
   sendNewMeetingSummary(meeting, context).catch(Logger.log)
   updateQualAIMeetingsCount(meeting.id, meeting.teamId, dataLoader)
@@ -50,13 +54,14 @@ const safeEndTeamPrompt = async ({
   context,
   subOptions
 }: {
-  meeting: MeetingTeamPrompt
+  meeting: TeamPromptMeeting
   now: Date
   viewerId?: string
   r: ParabolR
   context: InternalContext
   subOptions: SubOptions
 }) => {
+  const pg = getKysely()
   const {dataLoader} = context
 
   const {endedAt, id: meetingId, teamId} = meeting
@@ -65,7 +70,16 @@ const safeEndTeamPrompt = async ({
 
   // RESOLUTION
   const insights = await gatherInsights(meeting, dataLoader)
-  const completedTeamPrompt = (await r
+  await pg
+    .updateTable('NewMeeting')
+    .set({
+      endedAt: sql`CURRENT_TIMESTAMP`,
+      usedReactjis: JSON.stringify(insights.usedReactjis),
+      engagement: insights.engagement
+    })
+    .where('id', '=', meetingId)
+    .execute()
+  const completedTeamPrompt = await r
     .table('NewMeeting')
     .get(meetingId)
     .update(
@@ -76,12 +90,16 @@ const safeEndTeamPrompt = async ({
       {returnChanges: true}
     )('changes')(0)('new_val')
     .default(null)
-    .run()) as unknown as MeetingTeamPrompt
+    .run()
 
   if (!completedTeamPrompt) {
     return standardError(new Error('Completed team prompt meeting does not exist'), {
       userId: viewerId
     })
+  }
+
+  if (completedTeamPrompt.meetingType !== 'teamPrompt') {
+    return standardError(new Error('Meeting is not a team prompt'), {userId: viewerId})
   }
 
   const [meetingMembers, team, teamMembers, responses] = await Promise.all([
@@ -101,7 +119,6 @@ const safeEndTeamPrompt = async ({
       })
   )
   const timelineEventId = events[0]!.id
-  const pg = getKysely()
   await pg.insertInto('TimelineEvent').values(events).execute()
   summarizeTeamPrompt(meeting, context)
   analytics.teamPromptEnd(completedTeamPrompt, meetingMembers, responses, dataLoader)