+++ packages/server/graphql/mutations/helpers/sendPokerMeetingRevoteEvent.ts
@@ -1,11 +1,11 @@
-import Meeting from '../../../database/types/Meeting'
 import MeetingMember from '../../../database/types/MeetingMember'
 import {TeamMember} from '../../../postgres/types'
+import {AnyMeeting} from '../../../postgres/types/Meeting'
 import {analytics} from '../../../utils/analytics/analytics'
 import {DataLoaderWorker} from '../../graphql'
 
 const sendPokerMeetingRevoteEvent = async (
-  meeting: Meeting,
+  meeting: AnyMeeting,
   teamMembers: TeamMember[],
   meetingMembers: MeetingMember[],
   dataLoader: DataLoaderWorker