+++ packages/server/graphql/mutations/startSprintPoker.ts
@@ -8,7 +8,8 @@ import generateUID from '../../generateUID'
 import getKysely from '../../postgres/getKysely'
 import updateMeetingTemplateLastUsedAt from '../../postgres/queries/updateMeetingTemplateLastUsedAt'
 import updateTeamByTeamId from '../../postgres/queries/updateTeamByTeamId'
-import {MeetingTypeEnum} from '../../postgres/types/Meeting'
+import {MeetingTypeEnum, PokerMeeting} from '../../postgres/types/Meeting'
+import {PokerMeetingPhase} from '../../postgres/types/NewMeetingPhase'
 import {analytics} from '../../utils/analytics/analytics'
 import {getUserId, isTeamMember} from '../../utils/authorization'
 import getHashAndJSON from '../../utils/getHashAndJSON'
@@ -96,6 +97,7 @@ export default {
     }: {teamId: string; name: string | null | undefined; gcalInput?: CreateGcalEventInputType},
     {authToken, socketId: mutatorId, dataLoader}: GQLContext
   ) {
+    const pg = getKysely()
     const r = await getRethink()
     const operationId = dataLoader.share()
     const subOptions = {mutatorId, operationId}
@@ -115,15 +117,9 @@ export default {
 
     // RESOLUTION
     const meetingId = generateUID()
-    const meetingCount = await r
-      .table('NewMeeting')
-      .getAll(teamId, {index: 'teamId'})
-      .filter({meetingType})
-      .count()
-      .default(0)
-      .run()
+    const meetingCount = await dataLoader.get('meetingCount').load({teamId, meetingType})
 
-    const phases = await createNewMeetingPhases(
+    const phases = await createNewMeetingPhases<PokerMeetingPhase>(
       viewerId,
       teamId,
       meetingId,
@@ -149,14 +145,18 @@ export default {
       facilitatorUserId: viewerId,
       templateId: selectedTemplateId,
       templateRefId
-    })
+    }) as PokerMeeting
 
     const template = await dataLoader.get('meetingTemplates').load(selectedTemplateId)
-    await Promise.all([
+    await Promise.allSettled([
+      pg
+        .insertInto('NewMeeting')
+        .values({...meeting, phases: JSON.stringify(phases)})
+        .execute(),
       r.table('NewMeeting').insert(meeting).run(),
       updateMeetingTemplateLastUsedAt(selectedTemplateId, teamId)
     ])
-
+    dataLoader.clearAll('newMeetings')
     // Disallow accidental starts (2 meetings within 2 seconds)
     const newActiveMeetings = await dataLoader.get('activeMeetingsByTeamId').load(teamId)
     const otherActiveMeeting = newActiveMeetings.find((activeMeeting) => {