+++ packages/server/graphql/mutations/updatePokerScope.ts
@@ -4,7 +4,6 @@ import {SubscriptionChannel, Threshold} from 'parabol-client/types/constEnums'
 import {ESTIMATE_TASK_SORT_ORDER} from '../../../client/utils/constants'
 import getRethink from '../../database/rethinkDriver'
 import EstimateStage from '../../database/types/EstimateStage'
-import MeetingPoker from '../../database/types/MeetingPoker'
 import {TaskServiceEnum} from '../../database/types/Task'
 import getKysely from '../../postgres/getKysely'
 import {Discussion} from '../../postgres/pg'
@@ -43,6 +42,7 @@ const updatePokerScope = {
     {meetingId, updates}: {meetingId: string; updates: TUpdatePokerScopeItemInput[]},
     {authToken, dataLoader, socketId: mutatorId}: GQLContext
   ) => {
+    const pg = getKysely()
     const r = await getRethink()
     const redis = getRedis()
     const viewerId = getUserId(authToken)
@@ -56,22 +56,20 @@ const updatePokerScope = {
     // Wrap everything in try catch to ensure the lock is released
     try {
       //AUTH
-      const meeting = (await dataLoader.get('newMeetings').load(meetingId)) as MeetingPoker
+      const meeting = await dataLoader.get('newMeetings').load(meetingId)
       if (!meeting) {
         return {error: {message: `Meeting not found`}}
       }
-
-      const {endedAt, teamId, phases, meetingType, templateRefId, facilitatorStageId} = meeting
+      if (meeting.meetingType !== 'poker') {
+        return {error: {message: 'Not a poker meeting'}}
+      }
+      const {endedAt, teamId, phases, templateRefId, facilitatorStageId} = meeting
       if (!isTeamMember(authToken, teamId)) {
         // bad actors could be naughty & just lock meetings that they don't own. Limit bad actors to team members
         return {error: {message: `Not on team`}}
       }
       if (endedAt) {
         return {error: {message: `Meeting already ended`}}
-      }
-
-      if (meetingType !== 'poker') {
-        return {error: {message: 'Not a poker meeting'}}
       }
 
       // RESOLUTION
@@ -159,6 +157,15 @@ const updatePokerScope = {
       if (stages.length > Threshold.MAX_POKER_STORIES * dimensions.length) {
         return {error: {message: 'Story limit reached'}}
       }
+
+      await pg
+        .updateTable('NewMeeting')
+        .set({
+          facilitatorStageId: meeting.facilitatorStageId,
+          phases: JSON.stringify(phases)
+        })
+        .where('id', '=', meetingId)
+        .execute()
       await r
         .table('NewMeeting')
         .get(meetingId)
@@ -171,6 +178,7 @@ const updatePokerScope = {
       if (newDiscussions.length > 0) {
         await getKysely().insertInto('Discussion').values(newDiscussions).execute()
       }
+      dataLoader.clearAll(['newMeetings'])
       const data = {meetingId, newStageIds}
       publish(SubscriptionChannel.MEETING, meetingId, 'UpdatePokerScopeSuccess', data, subOptions)
       return data