+++ packages/server/graphql/public/types/AddTranscriptionBotSuccess.ts
@@ -1,4 +1,3 @@
-import MeetingRetrospective from '../../../database/types/MeetingRetrospective'
 import {AddTranscriptionBotSuccessResolvers} from '../resolverTypes'
 
 export type AddTranscriptionBotSuccessSource = {
@@ -8,7 +7,9 @@ export type AddTranscriptionBotSuccessSource = {
 const AddTranscriptionBotSuccess: AddTranscriptionBotSuccessResolvers = {
   meeting: async ({meetingId}, _args, {dataLoader}) => {
     const meeting = await dataLoader.get('newMeetings').load(meetingId)
-    return meeting as MeetingRetrospective
+    if (meeting.meetingType !== 'retrospective')
+      throw new Error('Meeting type is not retrospective')
+    return meeting
   }
 }