+++ packages/server/graphql/public/types/UpdateDimensionFieldSuccess.ts
@@ -1,4 +1,3 @@
-import MeetingPoker from '../../../database/types/MeetingPoker'
 import {UpdateDimensionFieldSuccessResolvers} from '../resolverTypes'
 
 export type UpdateDimensionFieldSuccessSource = {
@@ -10,7 +9,8 @@ const UpdateDimensionFieldSuccess: UpdateDimensionFieldSuccessResolvers = {
   team: ({teamId}, _args, {dataLoader}) => dataLoader.get('teams').loadNonNull(teamId),
   meeting: async ({meetingId}, _args, {dataLoader}) => {
     const meeting = await dataLoader.get('newMeetings').load(meetingId)
-    return meeting as MeetingPoker
+    if (meeting.meetingType !== 'poker') throw new Error('Not a poker meeting')
+    return meeting
   }
 }